Data on countries and their main city or agglomeration

There are firstly three identification codes of the country according to the ISO classification, the country's area in square kilometers, used to calculate in particular its internal distance. Variables indicating whether the country is landlocked and which continent it is part of are also included.

Format

A data frame with 238 observations on the following 34 variables.

iso2

ISO codes in two characters.

iso3

ISO codes in three characters.

cnum

ISO codes in three numbers.

country

Name of country in English.

pays

Name of country in French.

area

Country's area in km2.

dis_int

Internal distance of country i, dii=.67*sqrt(area/pi) (an often used measure of average distance between producers and consumers in a country). See Head and Mayer, 2002 for more on this topic.

landlocked

Dummy variable set equal to 1 for landlocked countries.

continent

Continent to which the country is belonging.

city_en

Names of capitals or main cities of the country in English.

city_fr

Names of capitals or main cities of the country in French.

lat

Latitude of the city.

lon

Longitude of the city.

cap

Variable equals to 1 if the city is the capital of the country, to 0 if the city is the most populated city (maincity equals to 1) but not the capital, and to 2 in the cases of two capitals, if the city is the most populated but the "second" capital or the previous capital.

maincity

Variable coded as 1 when the city is the most populated of the country and as 2 otherwise.

citynum

Number of cities for each country used to calculate the weighted distances described in Mayer and Zignago, 2011.

langoff_1

Official or national languages and languages spoken by at least 20% of the population of the country (and spoken in another country of the world) following the same logic than the "open-circuit languages" in Mélitz (2002).

langoff_2

Same as langoff_1.

langoff_3

Same as langoff_1.

lang20_1

Languages (mother tongue, lingua francas or second languages) spoken by at least 20% of the population of the country.

lang20_2

Same as lang20_1.

lang20_3

Same as lang20_1.

lang20_4

Same as lang20_1.

lang9_1

Languages (mother tongue, lingua francas or second languages) spoken by between 9% amd 20% of the population of the country.

lang9_2

Same as lang9_1.

lang9_3

Same as lang9_1.

lang9_4

Same as lang9_1.

colonizer1

Colonizers of the country for a relatively long period of time and with asubstantial participation in the governance of the colonized country.

colonizer2

Same as colonizer1.

colonizer3

Same as colonizer1.

colonizer4

Same as colonizer1.

short_colonizer1

Colonizers of the country for a relatively short period of time orwith only low involvement in the governance of the colonized country.

short_colonizer2

Same as short_colonizer1.

short_colonizer3

Same as short_colonizer1.

Source

http://www.cepii.fr/CEPII/en/bdd_modele/download.asp?id=6

References

Mayer, T. & Zignago, S. (2011) Notes on CEPII's distances measures: the GeoDist Database CEPII Working Paper 2011-25

Head, K. & Mayer, T. (2002) Illusory Border Effects: Distance Mismeasurement In-flates Estimates of Home Bias in Trade CEPII Working Paper 2002-01

Examples

# filter to avoid multiple records for the same country
geo_cepii[geo_cepii$cap == 1 & geo_cepii$maincity == 1, ]

#> # A tibble: 212 x 34
#>    iso2  iso3   cnum country pays    area dis_int landlocked continent city_en
#>    <chr> <chr> <dbl> <chr>   <chr>  <dbl>   <dbl>      <dbl> <chr>     <chr>  
#>  1 AW    ABW     533 Aruba   Aruba 1.93e2    5.23          0 America   Oranje~
#>  2 AF    AFG       4 Afghan~ Afgh~ 6.52e5  304.            1 Asia      Kabul  
#>  3 AO    AGO      24 Angola  Ango~ 1.25e6  420.            0 Africa    Luanda 
#>  4 AI    AIA     660 Anguil~ Angu~ 1.02e2    3.80          0 America   The Va~
#>  5 AL    ALB       8 Albania Alba~ 2.87e4   63.8           0 Europe    Tirana 
#>  6 AD    AND      20 Andorra Ando~ 4.53e2    8.01          0 Europe    Andorr~
#>  7 AN    ANT     530 Nether~ Anti~ 9.60e2   11.7           0 America   Willem~
#>  8 AE    ARE     784 United~ Emir~ 8.37e4  109.            0 Asia      Abu Dh~
#>  9 AR    ARG      32 Argent~ Arge~ 2.77e6  626.            0 America   Buenos~
#> 10 AM    ARM      51 Armenia Armé~ 2.98e4   64.9           1 Asia      Yerevan
#> # ... with 202 more rows, and 24 more variables: city_fr <chr>, lat <dbl>,
#> #   lon <dbl>, cap <dbl>, maincity <dbl>, citynum <chr>, langoff_1 <chr>,
#> #   langoff_2 <chr>, langoff_3 <chr>, lang20_1 <chr>, lang20_2 <chr>,
#> #   lang20_3 <chr>, lang20_4 <chr>, lang9_1 <chr>, lang9_2 <chr>,
#> #   lang9_3 <chr>, lang9_4 <chr>, colonizer1 <chr>, colonizer2 <chr>,
#> #   colonizer3 <chr>, colonizer4 <chr>, short_colonizer1 <chr>,
#> #   short_colonizer2 <chr>, short_colonizer3 <chr>